requirements.txt
# Terminal Agent Requirements
# Python 3.10+

# OpenAI SDK (LLM, Whisper STT fallback, GPT-4o-mini TTS)
openai>=1.40.0

# Groq SDK (ultra-fast Whisper Large V3 Turbo STT - 5-10x faster!)
groq>=0.4.0

# Environment management
python-dotenv>=1.0.0

# Audio recording/playback and file I/O
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# HTTP client for REST fallbacks
requests>=2.31.0

# Memory with automatic entity extraction and graph support
mem0ai>=0.0.1

# Vector database client (used by mem0)
qdrant-client>=1.7.0

# Graph database client (used by mem0)
neo4j>=5.20.0
langchain-neo4j
rank-bm25

# Twilio Voice Bridge
flask>=3.0.0
flask-sock>=0.7.0

# Telegram Voice Bot
python-telegram-bot>=20.0

# Discord Voice Bot
discord.py[voice]>=2.3.0
PyNaCl>=1.5.0

# WebRTC Voice Bridge (already have Flask from Twilio bridge)
# flask>=3.0.0 (already included above)
# flask-sock>=0.7.0 (already included above)

# LiveKit for ultra-fast real-time streaming (optional - for advanced deployment)
# livekit>=0.10.0
# livekit-agents>=0.8.0
# livekit-plugins-openai>=0.6.0
# livekit-plugins-silero>=0.6.0

# LangChain for document loading and advanced retrieval
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-groq>=0.1.0
langchain-qdrant>=0.1.0

# Document loaders
pypdf>=4.0.0
unstructured>=0.12.0
python-magic-bin>=0.4.14
openpyxl>=3.1.0